% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combine_cut}
\alias{combine_cut}
\title{Combine cut results}
\usage{
combine_cut(cut_path_up, cut_path_down, cut_path_any, table_cut)
}
\arguments{
\item{cut_path_up}{The cut_path_up table}

\item{cut_path_down}{The cut_path_down table}

\item{cut_path_any}{The cut_path_any table}

\item{table_cut}{The table_cut parameter}
}
\description{
This function takes the cut_path_up, cut_path_down and
cut_path_any tables and returns the results tables.
}
\examples{
# Add example here
print(combine_cut)
}
