% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combine_df}
\alias{combine_df}
\title{Combine df_up and df_down}
\usage{
combine_df(df_up, df_down, df_any)
}
\arguments{
\item{df_up}{The df_up table}

\item{df_down}{The df_down table}

\item{df_any}{The df_any table}
}
\description{
This function combines the results of the df_up and df_down tables.
}
\examples{
# Add examples here
print(combine_df)
}
