% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebba.R
\name{cutoff_path}
\alias{cutoff_path}
\title{Cutoff pathway}
\usage{
cutoff_path(path_table, p_cut, direction)
}
\arguments{
\item{path_table}{A table with some kind of information.}

\item{p_cut}{P-value cut.}

\item{direction}{The direction.}
}
\description{
This function takes a table and returns a dataframe with
several different types of information about pathways.
}
