% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Sepsis differential gene expression data as given by LIMMA for GEO study GSE49757}
\format{An object of class \code{data.frame}}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE49757}{GEO}
}
\usage{
data(example_data)
}
\description{
Results of a LIMMA differential gene expression analysis for a sepsis study by Khaenam et al, 2013.
}
\examples{
data(example_data)
head(example_data)
}
\keyword{datasets}
