% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{export_data}
\alias{export_data}
\title{Export data}
\usage{
export_data(file_in, df_combined, df_c, path_up, path_down, path_any,
  analysis_name, results_dir)
}
\arguments{
\item{file_in}{The input file}

\item{df_combined}{The result of the combine_cut function}

\item{df_c}{The result of the combine_df function}

\item{path_up}{Up pathways}

\item{path_down}{Down pathways}

\item{path_any}{Any pathways}

\item{analysis_name}{The name to be given to the analysis (Default: "Analysis1").}

\item{results_dir}{The folder into which results should be saved. A subdirectory
called "Heatmaps" will be created inside this folder.}
}
\description{
This function takes the results of the analyses and returns txt files.
}
\examples{
print(export_data)
}
