% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{faircc}
\alias{faircc}
\title{FAIRcc analysis}
\usage{
faircc(file_in, gmt_file, gene_col = "Gene.symbol", logFC_col = "logFC",
  pvalue_col = "P.Value", top_n = 3000, min_genes = 50,
  max_genes = 3000, p_cut = 0.2, order_p = "first", verbose = TRUE,
  analysis_name = NULL, results_dir = "Results", force = FALSE)
}
\arguments{
\item{file_in}{The file to execute the analysis}

\item{gmt_file}{The name of the gmt file containing terms and genes}

\item{gene_col}{A string indicating the column with genes (Default: "Gene.symbol")}

\item{logFC_col}{A string indicating the column with log fold-change
values (Default: "logFC").}

\item{pvalue_col}{A string indicating the column with p-values (Default: "P.Value")}

\item{top_n}{Top number of genes (Default: 3000)}

\item{min_genes}{Minimum number of genes (Default: 50)}

\item{max_genes}{Maximum number of genes (Default: 3000)}

\item{p_cut}{P-value cutoff (Default: 0.2)}

\item{order_p}{P-value ordering (Default:"first")}

\item{verbose}{Logical. If TRUE (default), will display analysis progress messages.}

\item{analysis_name}{The name to give to analysis results.
(Defaults to parameter \code{file_in} without the extensions and file path)}

\item{results_dir}{The path into which results should be saved (Default: "Results").}

\item{force}{Whether or not to overrwrite an existing results directory (Default: FALSE).}
}
\description{
This function executes the FAIRcc analysis.
}
\examples{
\dontrun{
# Run FAIRcc analyses
file_in <- "GSE49757_Septic_vs_Healthy.txt"
gmt_file <- "c2.cp.reactome.v6.0.symbols(1).gmt"
faircc(file_in, gmt_file)
}
}
