% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebba.R
\name{get_pathway}
\alias{get_pathway}
\title{Get pathway}
\usage{
get_pathway(merge_p, term2gene, all_genes, deg_list, gene_col, logFC_col,
  pvalue_col, direction, top_n = 3000, min_genes = 50, max_genes = 3000,
  p_cut = 0.01, order_p = "first")
}
\arguments{
\item{merge_p}{Something to merge}

\item{term2gene}{A data frame containing genes and terms}

\item{all_genes}{An object with all genes}

\item{deg_list}{A list with DEGs}

\item{gene_col}{A string indicating the column with genes}

\item{logFC_col}{A string indicating the column with log fold-change values}

\item{pvalue_col}{A string indicating the column with p-values}

\item{direction}{The direction. One of "up", "down" or "any"}

\item{top_n}{Top number of genes}

\item{min_genes}{Minimum number of genes}

\item{max_genes}{Maximum number of genes}

\item{p_cut}{P-value cut}

\item{order_p}{P-value ordering}
}
\description{
This function gets pathways.
}
\examples{
# Add examples here
print(get_pathway)
}
