% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebba.R
\name{run_enrich}
\alias{run_enrich}
\title{Run Enricher}
\usage{
run_enrich(top_genes, all_genes, term2gene)
}
\arguments{
\item{top_genes}{A vector of genes}

\item{all_genes}{An object with all genes}

\item{term2gene}{A data.frame with enrichment term and genes}
}
\description{
This function runs enricher
}
\examples{
# Add example here
print(run_enrich)
}
